pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "llmstitch"
version = "0.1.0"
description = "A provider-agnostic LLM toolkit with tool calling, skills, and parallel execution."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Biniam Kassahun", email = "george.beng@gmail.com" },
]
keywords = ["llm", "anthropic", "openai", "gemini", "tool-calling", "agents", "ai"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Typing :: Typed",
]
dependencies = []

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai    = ["openai>=1.50"]
gemini    = ["google-genai>=0.3"]
all       = ["llmstitch[anthropic,openai,gemini]"]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
  "ruff>=0.6",
  "mypy>=1.11",
  "pre-commit>=3",
]

[project.urls]
Homepage   = "https://github.com/bengeos/llmstitch"
Repository = "https://github.com/bengeos/llmstitch"
Issues     = "https://github.com/bengeos/llmstitch/issues"
Changelog  = "https://github.com/bengeos/llmstitch/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/llmstitch"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
# Vendor SDKs are optional extras; don't hard-fail when they're not installed.
ignore_missing_imports = true
